+++ dflib/src/test/java/org/dflib/DataFrame_GetColumnTest.java
@@ -22,6 +22,12 @@ public class DataFrame_GetColumnTest {
     }
 
     @Test
+    public void byLabelException() {
+        DataFrame df = DataFrame.foldByRow("a").of(1, 2);
+        assertThrows(IllegalArgumentException.class, () -> df.getColumn("x"));
+    }
+
+    @Test
     public void byPosition() {
         DataFrame df = DataFrame.foldByRow("a", "b").of(
                 1, "x",
@@ -30,6 +36,13 @@ public class DataFrame_GetColumnTest {
         Series<String> cb = df.getColumn(0);
 
         new SeriesAsserts(cb).expectData(1, 2);
+    }
+
+    @Test
+    public void byPositionException() {
+        DataFrame df = DataFrame.foldByRow("a").of(1, 2);
+        assertThrows(IllegalArgumentException.class, () -> df.getColumn(1));
+        assertThrows(IllegalArgumentException.class, () -> df.getColumn(-1));
     }
 
     @Deprecated